requirements.txt
jinja2
python-dotenv
fastapi
google-adk
anthropic[vertex]
litellm
pandas
numpy
matplotlib
num2words
vllm
huggingface_hub[cli]